pyproject.toml
[project]
name = "agentic-web-builder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "langchain>=1.2.0",
    "langchain-core>=1.2.0",
    "langchain-groq>=1.1.0",
    "langgraph>=1.1.0",
    "python-dotenv>=1.0.0"
]
